Split CsvPanel into focused csv-panel components for map tools, file controls, metadata, coordinate mapping, parsing warnings, and preview table.
Timeline controls intentionally remain in CsvPanel because shared timeline work is covered by #61.
